+++ src/main/java/com/ll/dopdang/domain/project/controller/ContractController.java
@@ -1,0 +1,48 @@
+package com.ll.dopdang.domain.project.controller;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.ll.dopdang.domain.project.dto.ContractCreateRequest;
+import com.ll.dopdang.domain.project.service.ContractService;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@RestController
+@RequestMapping("/contracts")
+@RequiredArgsConstructor
+@Slf4j
+public class ContractController {
+
+	private final ContractService contractService;
+
+	/**
+	 * 계약 생성 API
+	 * @param request 계약 생성 요청 정보
+	 * @return 생성된 계약 ID
+	 */
+	@PostMapping
+	public ResponseEntity<?> createContract(@RequestBody ContractCreateRequest request) {
+		log.info("계약 생성 요청: offerId={}, price={}, startDate={}, endDate={}",
+			request.getOfferId(), request.getPrice(), request.getStartDate(), request.getEndDate());
+
+		Long contractId = contractService.createContractFromOffer(
+			request.getOfferId(),
+			request.getPrice(),
+			request.getStartDate(),
+			request.getEndDate()
+		);
+
+		Map<String, Object> response = new HashMap<>();
+		response.put("contractId", contractId);
+
+		return ResponseEntity.ok(response);
+	}
+}
